refactored correlation to take advantage of dynamic sizes, added contingency for no fields by giving defaults to values, actually using workspace properly.

+++ runtime/cerberus_net/trt_plugins/grid_sampler.cpp
@@ -51,7 +51,7 @@ void GridSamplerPlugin::serialize(void* buffer) const
 {
     char* d = static_cast<char*>(buffer), *a = d;
 
-    write(d, static_cast<int>(m_align_corners));
+    write(d, m_align_corners);
     write(d, static_cast<int>(m_interpolation_mode));
     write(d, static_cast<int>(m_padding_mode));
 
@@ -62,7 +62,7 @@ size_t GridSamplerPlugin::getSerializationSize() const
 {
     size_t serializationSize = 0;
 
-    serializationSize += sizeof(static_cast<int>(m_align_corners));
+    serializationSize += sizeof(m_align_corners);
     serializationSize += sizeof(static_cast<int>(m_interpolation_mode));
     serializationSize += sizeof(static_cast<int>(m_padding_mode));
